refactor(WebMercatorExtents): add grid to world conversion methods
These were originally on Grid but they are potentially useful from any
object that has WebMercatorExtents.
