+++ src/main/java/com/conveyal/r5/analyst/Grid.java
@@ -7,7 +7,6 @@ import com.conveyal.r5.util.ShapefileReader;
 import com.csvreader.CsvReader;
 import com.google.common.io.LittleEndianDataInputStream;
 import com.google.common.io.LittleEndianDataOutputStream;
-import com.google.common.primitives.Doubles;
 import org.apache.commons.math3.util.FastMath;
 import org.geotools.coverage.grid.GridCoverage2D;
 import org.geotools.coverage.grid.GridCoverageFactory;
@@ -22,7 +21,6 @@ import org.geotools.gce.geotiff.GeoTiffFormat;
 import org.geotools.gce.geotiff.GeoTiffWriteParams;
 import org.geotools.gce.geotiff.GeoTiffWriter;
 import org.geotools.geometry.jts.ReferencedEnvelope;
-import org.geotools.referencing.CRS;
 import org.geotools.referencing.crs.DefaultGeographicCRS;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
@@ -39,7 +37,6 @@ import org.opengis.feature.simple.SimpleFeatureType;
 import org.opengis.parameter.GeneralParameterValue;
 import org.opengis.parameter.ParameterValueGroup;
 import org.opengis.referencing.FactoryException;
-import org.opengis.referencing.crs.CoordinateReferenceSystem;
 import org.opengis.referencing.operation.TransformException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -323,45 +320,6 @@ public class Grid extends PointSet {
         out.close();
     }
 
-    /**
-     * How to get the width of the world in meters according to the EPSG CRS spec:
-     * $ gdaltransform -s_srs epsg:4326 -t_srs epsg:3857
-     * 180, 0
-     * 20037508.3427892 -7.08115455161362e-10 0
-     * You can't do 180, 90 because this projection is cut off above a certain level to make the world square.
-     * You can do the reverse projection to find this latitude:
-     * $ gdaltransform -s_srs epsg:3857 -t_srs epsg:4326
-     * 20037508.342789, 20037508.342789
-     * 179.999999999998 85.0511287798064 0
-     */
-    public Coordinate mercatorPixelToMeters (double xPixel, double yPixel) {
-        double worldWidthPixels = Math.pow(2, zoom) * 256D;
-        // Top left is min x and y because y increases toward the south in web Mercator. Bottom right is max x and y.
-        // The origin is WGS84 (0,0).
-        final double worldWidthMeters = 20037508.342789244 * 2;
-        double xMeters = ((xPixel / worldWidthPixels) - 0.5) * worldWidthMeters;
-        double yMeters = (0.5 - (yPixel / worldWidthPixels)) * worldWidthMeters; // flip y axis
-        return new Coordinate(xMeters, yMeters);
-    }
-
-    /**
-     * At zoom level zero, our coordinates are pixels in a single planetary tile, with coordinates are in the range
-     * [0...256). We want to export with a conventional web Mercator envelope in meters.
-     */
-    public ReferencedEnvelope getMercatorEnvelopeMeters() {
-        Coordinate topLeft = mercatorPixelToMeters(west, north);
-        Coordinate bottomRight = mercatorPixelToMeters(west + width, north + height);
-        Envelope mercatorEnvelope = new Envelope(topLeft, bottomRight);
-        try {
-            // Get Spherical Mercator pseudo-projection CRS
-            CoordinateReferenceSystem webMercator = CRS.decode("EPSG:3857");
-            ReferencedEnvelope env = new ReferencedEnvelope(mercatorEnvelope, webMercator);
-            return env;
-        } catch (Exception ex) {
-            throw new RuntimeException(ex);
-        }
-    }
-
     /** Write this grid out in GeoTIFF format */
     public void writeGeotiff (OutputStream out) {
         try {
@@ -371,7 +329,7 @@ public class Grid extends PointSet {
                     data[y][x] = (float) grid[x][y];
                 }
             }
-            ReferencedEnvelope env = getMercatorEnvelopeMeters();
+            ReferencedEnvelope env = this.getWebMercatorExtents().getMercatorEnvelopeMeters();
             GridCoverage2D coverage = new GridCoverageFactory().create("GRID", data, env);
             GeoTiffWriteParams wp = new GeoTiffWriteParams();
             wp.setCompressionMode(GeoTiffWriteParams.MODE_EXPLICIT);
